pyproject.toml
# Copyright (c) 2026 TRIBUNAL Team.
# Smart Factory Assembly — OpenEnv Hackathon Submission.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-smart-factory"
version = "1.0.0"
description = "Smart Factory Assembly Environment for OpenEnv — Multi-difficulty manufacturing RL"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
inference = [
    "openai>=1.0.0",
    "requests>=2.31.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["smart_factory", "smart_factory.server"]
package-dir = { "smart_factory" = ".", "smart_factory.server" = "server" }